requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
tensorflow==2.13.0
joblib==1.3.2
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0 